requirements.txt
discord.py==2.3.2
selenium==4.21.0